finish excercise view and view group
