+++ app/build.gradle
@@ -63,5 +63,4 @@ dependencies {
     implementation 'de.hdodenhof:circleimageview:2.2.0'
     implementation 'com.android.support:design:27.1.1'
     apply from: "../ci-java.gradle"
-
 }
