+++ app/src/at-thachnguyen/AndroidManifest.xml
@@ -9,15 +9,14 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".thachnguyen.debug.MenuActivity">
+        <activity android:name=".thachnguyen.debug.menu.MenuActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".thachnguyen.debug.Main2Activity">
+        <activity android:name=".thachnguyen.debug.view_and_view_group.Main2Activity">
         </activity>
     </application>
-
 </manifest>
