+++ app/src/at-thachnguyen/res/drawable/radius_view.xml
@@ -1,0 +1,6 @@
+<shape
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <corners android:radius="15dp" />
+    <solid android:color="@color/colorTurquoiseBlue" />
+</shape>
