+++ app/src/at-thachnguyen/res/drawable/radius_view_day.xml
@@ -1,0 +1,6 @@
+<shape
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <corners android:radius="40dp" />
+    <solid android:color="@color/colorTimberwolfapprox" />
+</shape>
