+++ app/src/at-thachnguyen/res/layout/activity_menu.xml
@@ -3,28 +3,28 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".thachnguyen.debug.MenuActivity"
-    android:orientation="vertical">
-<Button
-    android:background="#109b4a"
-    android:layout_marginTop="20dp"
-    android:layout_marginLeft="20dp"
-    android:layout_marginRight="20dp"
-    android:id="@+id/btnEx1"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:text="@string/ex1"
-    android:textStyle="bold"
-    />
+    android:orientation="vertical"
+    tools:context=".thachnguyen.debug.menu.MenuActivity">
+
     <Button
-        android:background="#e2be09"
-        android:layout_marginTop="20dp"
+        android:id="@+id/btnEx1"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
         android:layout_marginLeft="20dp"
         android:layout_marginRight="20dp"
+        android:layout_marginTop="20dp"
+        android:background="@color/colorSalemapprox"
+        android:text="@string/ex1"
+        android:textStyle="bold" />
+
+    <Button
         android:id="@+id/btnEx2"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
+        android:layout_marginLeft="20dp"
+        android:layout_marginRight="20dp"
+        android:layout_marginTop="20dp"
+        android:background="@color/colorCornapprox"
         android:text="@string/ex2"
-        android:textStyle="bold"
-        />
+        android:textStyle="bold" />
 </LinearLayout>
