+++ app/src/at-thachnguyen/res/values/colors.xml
@@ -5,4 +5,9 @@
     <color name="colorAccent">#FF4081</color>
     <color name="colorWhite">#FFFFFF</color>
     <color name="colorGreen">#008000</color>
+    <color name="colorBlack">#000000</color>
+    <color name="colorTurquoiseBlue">#61EDC0</color>
+    <color name="colorCornapprox">#E2BE09</color>
+    <color name="colorSalemapprox">#109b4a</color>
+    <color name="colorTimberwolfapprox">#D8D7D0</color>
 </resources>
