+++ app/src/at-thachnguyen/res/values/strings.xml
@@ -7,6 +7,8 @@
     <string name="thu">THU</string>
     <string name="fri">FRI</string>
     <string name="sat">SAT</string>
+    <string name="february">FEBRUARY</string>
+    <string name="good_morning">Good Morning</string>
     <string name="_1">1</string>
     <string name="_2">2</string>
     <string name="_3">3</string>
@@ -23,4 +25,3 @@
     <string name="ex1">View and View Group</string>
     <string name="ex2">Intent</string>
 </resources>
-
